Move TrackSelectionList ui stuff to the list package
